requirements.txt
absl-py==0.10.0
anyio==3.2.1
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
docker==4.3.1
entrypoints==0.3
finnhub-python==2.4.3
fire==0.3.1
Flask==1.1.2
gast==0.3.3
google-auth==1.4.2
google-auth-oauthlib==0.4.1
google-colab==1.0.0
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.7.0
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
matplotlib-inline==0.1.2
mido==1.2.9
mistune==0.8.4
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==5.2.2
numpy==1.18.5
oauthlib==3.1.0
opam==1.0.6
openai==0.2.4
opt-einsum==3.3.0
packaging==21.0
pandas==0.24.2
pandocfilters==1.4.3
parso==0.8.2
patiencediff==0.2.2
pickleshare==0.7.5
Pillow==7.2.0
portpicker==1.2.0
praw==7.1.0
prawcore==1.5.0
pretty-midi==0.2.9
prometheus-client==0.11.0
prompt-toolkit==1.0.18
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pynput==1.7.3
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
pywin32==227
pywinpty==1.1.3
PyYAML==5.3.1
pyzmq==22.1.0
regex==2017.4.5
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.6
scipy==1.4.1
Send2Trash==1.7.1
simplegeneric==0.8.1
six==1.12.0
sniffio==1.2.0
soupsieve==2.2.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
tornado==4.5.3
tqdm==4.31.1
traitlets==5.0.5
typing-extensions==3.10.0.0
update-checker==0.18.0
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
